# cli/src/pcluster/api/openapi/openapi.yaml
openapi: 3.0.2
info:
  description: ParallelCluster API
  title: ParallelCluster
  version: 3.11.0
servers:
- url: /
security:
- aws.auth.sigv4: []
paths:
...
  /v3/images/custom:
    get:
      description: Retrieve the list of existing custom images.
      operationId: list_images
      parameters:
      - description: List images built in a given AWS Region.
        explode: true
        in: query
        name: region
        required: false
        schema:
          description: List images built in a given AWS Region.
          type: string
        style: form
      - description: Token to use for paginated requests.
        explode: true
        in: query
        name: nextToken
        required: false
        schema:
          description: Token to use for paginated requests.
          type: string
        style: form
      - description: Filter images by the status provided.
        explode: true
        in: query
        name: imageStatus
        required: true
        schema:
          $ref: '#/components/schemas/ImageStatusFilteringOption'
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListImagesResponseContent'
          description: ListImages 200 response
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestExceptionResponseContent'
          description: BadRequestException 400 response
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedClientErrorResponseContent'
          description: UnauthorizedClientError 401 response
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LimitExceededExceptionResponseContent'
          description: LimitExceededException 429 response
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServiceExceptionResponseContent'
          description: InternalServiceException 500 response
      tags:
      - Image Operations
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri:
          Fn::Sub: "arn:${AWS::Partition}:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${ParallelClusterFunction.Arn}/invocations"
        credentials:
          Fn::Sub: "${APIGatewayExecutionRole.Arn}"
        payloadFormatVersion: "2.0"
      x-openapi-router-controller: pcluster.api.controllers.image_operations_controller
    post:
      description: Create a custom ParallelCluster image in a given region.
      operationId: build_image
      parameters:
      - description: "Identifies one or more config validators to suppress. Format:\
          \ (ALL|type:[A-Za-z0-9]+)"
        explode: true
        in: query
        name: suppressValidators
        required: false
        schema:
          description: "Identifies one or more config validators to suppress. Format:\
            \ (ALL|type:[A-Za-z0-9]+)"
          items:
            pattern: "^(ALL|type:[A-Za-z0-9]+)$"
            type: string
          type: array
          uniqueItems: true
        style: form
      - description: Min validation level that will cause the creation to fail. (Defaults
          to 'ERROR'.)
        explode: true
        in: query
        name: validationFailureLevel
        required: false
        schema:
          $ref: '#/components/schemas/ValidationLevel'
        style: form
      - description: Only perform request validation without creating any resource.
          It can be used to validate the image configuration. (Defaults to 'false'.)
        explode: true
        in: query
        name: dryrun
        required: false
        schema:
          description: Only perform request validation without creating any resource.
            It can be used to validate the image configuration. (Defaults to 'false'.)
          type: boolean
        style: form
      - description: "When set, will automatically initiate an image stack rollback\
          \ on failure. (Defaults to 'false'.)"
        explode: true
        in: query
        name: rollbackOnFailure
        required: false
        schema:
          description: "When set, will automatically initiate an image stack rollback\
            \ on failure. (Defaults to 'false'.)"
          type: boolean
        style: form
      - description: AWS Region that the operation corresponds to.
        explode: true
        in: query
        name: region
        required: false
        schema:
          description: AWS Region that the operation corresponds to.
          type: string
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuildImageRequestContent'
        required: true
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildImageResponseContent'
          description: BuildImage 202 response
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildImageBadRequestExceptionResponseContent'
          description: BuildImageBadRequestException 400 response
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedClientErrorResponseContent'
          description: UnauthorizedClientError 401 response
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictExceptionResponseContent'
          description: ConflictException 409 response
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DryrunOperationExceptionResponseContent'
          description: DryrunOperationException 412 response
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LimitExceededExceptionResponseContent'
          description: LimitExceededException 429 response
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServiceExceptionResponseContent'
          description: InternalServiceException 500 response
      tags:
      - Image Operations
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri:
          Fn::Sub: "arn:${AWS::Partition}:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${ParallelClusterFunction.Arn}/invocations"
        credentials:
          Fn::Sub: "${APIGatewayExecutionRole.Arn}"
        payloadFormatVersion: "2.0"
      x-openapi-router-controller: pcluster.api.controllers.image_operations_controller
  /v3/images/custom/{imageId}:
    delete:
      description: Initiate the deletion of the custom ParallelCluster image.
      operationId: delete_image
      parameters:
      - description: Id of the image.
        explode: false
        in: path
        name: imageId
        required: true
        schema:
          description: Id of the image.
          pattern: "^[a-zA-Z][a-zA-Z0-9-]+$"
          type: string
        style: simple
      - description: AWS Region that the operation corresponds to.
        explode: true
        in: query
        name: region
        required: false
        schema:
          description: AWS Region that the operation corresponds to.
          type: string
        style: form
      - description: Force deletion in case there are instances using the AMI or in
          case the AMI is shared. (Defaults to 'false'.)
        explode: true
        in: query
        name: force
        required: false
        schema:
          description: Force deletion in case there are instances using the AMI or
            in case the AMI is shared. (Defaults to 'false'.)
          type: boolean
        style: form
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteImageResponseContent'
          description: DeleteImage 202 response
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestExceptionResponseContent'
          description: BadRequestException 400 response
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedClientErrorResponseContent'
          description: UnauthorizedClientError 401 response
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundExceptionResponseContent'
          description: NotFoundException 404 response
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LimitExceededExceptionResponseContent'
          description: LimitExceededException 429 response
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServiceExceptionResponseContent'
          description: InternalServiceException 500 response
      tags:
      - Image Operations
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri:
          Fn::Sub: "arn:${AWS::Partition}:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${ParallelClusterFunction.Arn}/invocations"
        credentials:
          Fn::Sub: "${APIGatewayExecutionRole.Arn}"
        payloadFormatVersion: "2.0"
      x-openapi-router-controller: pcluster.api.controllers.image_operations_controller
    get:
      description: Get detailed information about an existing image.
      operationId: describe_image
      parameters:
      - description: Id of the image.
        explode: false
        in: path
        name: imageId
        required: true
        schema:
          description: Id of the image.
          pattern: "^[a-zA-Z][a-zA-Z0-9-]+$"
          type: string
        style: simple
      - description: AWS Region that the operation corresponds to.
        explode: true
        in: query
        name: region
        required: false
        schema:
          description: AWS Region that the operation corresponds to.
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DescribeImageResponseContent'
          description: DescribeImage 200 response
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestExceptionResponseContent'
          description: BadRequestException 400 response
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedClientErrorResponseContent'
          description: UnauthorizedClientError 401 response
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundExceptionResponseContent'
          description: NotFoundException 404 response
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LimitExceededExceptionResponseContent'
          description: LimitExceededException 429 response
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServiceExceptionResponseContent'
          description: InternalServiceException 500 response
      tags:
      - Image Operations
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri:
          Fn::Sub: "arn:${AWS::Partition}:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${ParallelClusterFunction.Arn}/invocations"
        credentials:
          Fn::Sub: "${APIGatewayExecutionRole.Arn}"
        payloadFormatVersion: "2.0"
      x-openapi-router-controller: pcluster.api.controllers.image_operations_controller
...
  /v3/images/official:
    get:
      description: List Official ParallelCluster AMIs.
      operationId: list_official_images
      parameters:
      - description: AWS Region that the operation corresponds to.
        explode: true
        in: query
        name: region
        required: false
        schema:
          description: AWS Region that the operation corresponds to.
          type: string
        style: form
      - description: Filter by OS distribution (Default is to not filter.)
        explode: true
        in: query
        name: os
        required: false
        schema:
          description: Filter by OS distribution (Default is to not filter.)
          type: string
        style: form
      - description: Filter by architecture (Default is to not filter.)
        explode: true
        in: query
        name: architecture
        required: false
        schema:
          description: Filter by architecture (Default is to not filter.)
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListOfficialImagesResponseContent'
          description: ListOfficialImages 200 response
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestExceptionResponseContent'
          description: BadRequestException 400 response
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedClientErrorResponseContent'
          description: UnauthorizedClientError 401 response
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LimitExceededExceptionResponseContent'
          description: LimitExceededException 429 response
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServiceExceptionResponseContent'
          description: InternalServiceException 500 response
      tags:
      - Image Operations
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri:
          Fn::Sub: "arn:${AWS::Partition}:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${ParallelClusterFunction.Arn}/invocations"
        credentials:
          Fn::Sub: "${APIGatewayExecutionRole.Arn}"
        payloadFormatVersion: "2.0"
      x-openapi-router-controller: pcluster.api.controllers.image_operations_controller
components:
  schemas:
...
    BadRequestExceptionResponseContent:
      description: This exception is thrown when a client calls an API with wrong
        parameters.
      properties:
        message:
          title: message
          type: string
      title: BadRequestExceptionResponseContent
      type: object
    BuildImageBadRequestExceptionResponseContent:
      description: This exception is thrown when a client calls the BuildImage API
        with an invalid request. This includes an error due to invalid image configuration.
      properties:
        message:
          title: message
          type: string
        configurationValidationErrors:
          items:
            $ref: '#/components/schemas/ConfigValidationMessage'
          title: configurationValidationErrors
          type: array
      title: BuildImageBadRequestExceptionResponseContent
      type: object
    BuildImageRequestContent:
      example:
        imageConfiguration: imageConfiguration
        imageId: imageId
      properties:
        imageConfiguration:
          description: Image configuration as a YAML document.
          title: imageConfiguration
          type: string
        imageId:
          description: Id of the Image that will be built.
          pattern: "^[a-zA-Z][a-zA-Z0-9-]+$"
          title: imageId
          type: string
      required:
      - imageConfiguration
      - imageId
      title: BuildImageRequestContent
      type: object
    BuildImageResponseContent:
      example:
        image:
          imageId: imageId
          imageBuildStatus: null
          cloudformationStackStatus: null
          cloudformationStackArn: cloudformationStackArn
          ec2AmiInfo:
            amiId: amiId
          region: region
          version: version
        validationMessages:
        - level: null
          id: id
          type: type
          message: message
        - level: null
          id: id
          type: type
          message: message
      properties:
        image:
          $ref: '#/components/schemas/ImageInfoSummary'
        validationMessages:
          description: List of messages collected during image config validation whose
            level is lower than the 'validationFailureLevel' set by the user.
          items:
            $ref: '#/components/schemas/ConfigValidationMessage'
          title: validationMessages
          type: array
      required:
      - image
      title: BuildImageResponseContent
      type: object
    Change:
      example:
        parameter: parameter
        requestedValue: requestedValue
        currentValue: currentValue
      properties:
        parameter:
          title: parameter
          type: string
        currentValue:
          title: currentValue
          type: string
        requestedValue:
          title: requestedValue
          type: string
      title: Change
      type: object
...
    CloudFormationStackStatus:
      enum:
      - CREATE_IN_PROGRESS
      - CREATE_FAILED
      - CREATE_COMPLETE
      - IMPORT_IN_PROGRESS
      - IMPORT_COMPLETE
      - IMPORT_ROLLBACK_IN_PROGRESS
      - IMPORT_ROLLBACK_FAILED
      - IMPORT_ROLLBACK_COMPLETE
      - ROLLBACK_IN_PROGRESS
      - ROLLBACK_FAILED
      - ROLLBACK_COMPLETE
      - DELETE_IN_PROGRESS
      - DELETE_FAILED
      - DELETE_COMPLETE
      - REVIEW_IN_PROGRESS
      - UPDATE_IN_PROGRESS
      - UPDATE_COMPLETE_CLEANUP_IN_PROGRESS
      - UPDATE_COMPLETE
      - UPDATE_FAILED
      - UPDATE_ROLLBACK_IN_PROGRESS
      - UPDATE_ROLLBACK_FAILED
      - UPDATE_ROLLBACK_COMPLETE_CLEANUP_IN_PROGRESS
      - UPDATE_ROLLBACK_COMPLETE
      title: CloudFormationStackStatus
      type: string
...
    ConfigValidationMessage:
      example:
        level: null
        id: id
        type: type
        message: message
      properties:
        id:
          description: Id of the validator.
          title: id
          type: string
        type:
          description: Type of the validator.
          title: type
          type: string
        level:
          $ref: '#/components/schemas/ValidationLevel'
        message:
          description: Validation message
          title: message
          type: string
      title: ConfigValidationMessage
      type: object
    ConflictExceptionResponseContent:
      description: This exception is thrown when a client request to create/modify
        content would result in a conflict.
      properties:
        message:
          title: message
          type: string
      title: ConflictExceptionResponseContent
      type: object
...
    DeleteImageResponseContent:
      example:
        image:
          imageId: imageId
          imageBuildStatus: null
          cloudformationStackStatus: null
          cloudformationStackArn: cloudformationStackArn
          ec2AmiInfo:
            amiId: amiId
          region: region
          version: version
      properties:
        image:
          $ref: '#/components/schemas/ImageInfoSummary'
      required:
      - image
      title: DeleteImageResponseContent
      type: object
...
    DescribeImageResponseContent:
      example:
        cloudformationStackStatusReason: cloudformationStackStatusReason
        imagebuilderImageStatusReason: imagebuilderImageStatusReason
        imageId: imageId
        cloudformationStackTags:
        - value: value
          key: key
        - value: value
          key: key
        imagebuilderImageStatus: null
        creationTime: 2000-01-23T04:56:07.000+00:00
        version: version
        imageConfiguration:
          url: url
        imageBuildStatus: null
        cloudformationStackStatus: null
        imageBuildLogsArn: imageBuildLogsArn
        cloudformationStackArn: cloudformationStackArn
        region: region
        cloudformationStackCreationTime: 2000-01-23T04:56:07.000+00:00
        ec2AmiInfo:
          amiName: amiName
          amiId: amiId
          description: description
          state: null
          tags:
          - value: value
            key: key
          - value: value
            key: key
          architecture: architecture
      properties:
        imageId:
          description: Id of the Image to retrieve detailed information for.
          title: imageId
          type: string
        region:
          description: AWS region where the image is created.
          title: region
          type: string
        version:
          description: ParallelCluster version used to build the image.
          title: version
          type: string
        imageBuildStatus:
          $ref: '#/components/schemas/ImageBuildStatus'
        imageBuildLogsArn:
          description: ARN of the logs for the image build process.
          title: imageBuildLogsArn
          type: string
        cloudformationStackStatus:
          $ref: '#/components/schemas/CloudFormationStackStatus'
        cloudformationStackStatusReason:
          description: Reason for the CloudFormation stack status.
          title: cloudformationStackStatusReason
          type: string
        cloudformationStackArn:
          description: ARN of the main CloudFormation stack.
          title: cloudformationStackArn
          type: string
        creationTime:
          description: Timestamp representing the image creation time.
          format: date-time
          title: creationTime
          type: string
        cloudformationStackCreationTime:
          description: Timestamp representing the CloudFormation stack creation time.
          format: date-time
          title: cloudformationStackCreationTime
          type: string
        cloudformationStackTags:
          description: Tags for the CloudFormation stack.
          items:
            $ref: '#/components/schemas/Tag'
          title: cloudformationStackTags
          type: array
        imageConfiguration:
          $ref: '#/components/schemas/ImageConfigurationStructure'
        imagebuilderImageStatus:
          $ref: '#/components/schemas/ImageBuilderImageStatus'
        imagebuilderImageStatusReason:
          description: Reason for the ImageBuilder Image status.
          title: imagebuilderImageStatusReason
          type: string
        ec2AmiInfo:
          $ref: '#/components/schemas/Ec2AmiInfo'
      required:
      - imageBuildStatus
      - imageConfiguration
      - imageId
      - region
      - version
      title: DescribeImageResponseContent
      type: object
    DryrunOperationExceptionResponseContent:
      description: Communicates that the operation would have succeeded without the
        dryrun flag.
      properties:
        message:
          title: message
          type: string
        changeSet:
          description: List of configuration changes requested by the operation.
          items:
            $ref: '#/components/schemas/Change'
          title: changeSet
          type: array
        validationMessages:
          description: List of messages collected during cluster config validation
            whose level is lower than the 'validationFailureLevel' set by the user.
          items:
            $ref: '#/components/schemas/ConfigValidationMessage'
          title: validationMessages
          type: array
      title: DryrunOperationExceptionResponseContent
      type: object
...
    ImageBuildStatus:
      enum:
      - BUILD_IN_PROGRESS
      - BUILD_FAILED
      - BUILD_COMPLETE
      - DELETE_IN_PROGRESS
      - DELETE_FAILED
      - DELETE_COMPLETE
      title: ImageBuildStatus
      type: string
...
    ImageConfigurationStructure:
      example:
        url: url
      properties:
        url:
          description: URL of the image configuration file.
          title: url
          type: string
      title: ImageConfigurationStructure
      type: object
    ImageInfoSummary:
      example:
        imageId: imageId
        imageBuildStatus: null
        cloudformationStackStatus: null
        cloudformationStackArn: cloudformationStackArn
        ec2AmiInfo:
          amiId: amiId
        region: region
        version: version
      properties:
        imageId:
          description: Id of the image.
          pattern: "^[a-zA-Z][a-zA-Z0-9-]+$"
          title: imageId
          type: string
        ec2AmiInfo:
          $ref: '#/components/schemas/Ec2AmiInfoSummary'
        region:
          description: AWS region where the image is built.
          title: region
          type: string
        version:
          description: ParallelCluster version used to build the image.
          title: version
          type: string
        cloudformationStackArn:
          description: ARN of the main CloudFormation stack.
          title: cloudformationStackArn
          type: string
        imageBuildStatus:
          $ref: '#/components/schemas/ImageBuildStatus'
        cloudformationStackStatus:
          $ref: '#/components/schemas/CloudFormationStackStatus'
      required:
      - imageBuildStatus
      - imageId
      - region
      - version
      title: ImageInfoSummary
      type: object
    ImageStatusFilteringOption:
      enum:
      - AVAILABLE
      - PENDING
      - FAILED
      title: ImageStatusFilteringOption
      type: string
...
    InternalServiceExceptionResponseContent:
      description: This exception is thrown on an unhandled service error.
      properties:
        message:
          title: message
          type: string
      title: InternalServiceExceptionResponseContent
      type: object
    LimitExceededExceptionResponseContent:
      description: The client is sending more than the allowed number of requests
        per unit of time.
      properties:
        message:
          title: message
          type: string
      title: LimitExceededExceptionResponseContent
      type: object
...
    ListImagesResponseContent:
      example:
        images:
        - imageId: imageId
          imageBuildStatus: null
          cloudformationStackStatus: null
          cloudformationStackArn: cloudformationStackArn
          ec2AmiInfo:
            amiId: amiId
          region: region
          version: version
        - imageId: imageId
          imageBuildStatus: null
          cloudformationStackStatus: null
          cloudformationStackArn: cloudformationStackArn
          ec2AmiInfo:
            amiId: amiId
          region: region
          version: version
        nextToken: nextToken
      properties:
        nextToken:
          description: Token to use for paginated requests.
          title: nextToken
          type: string
        images:
          items:
            $ref: '#/components/schemas/ImageInfoSummary'
          title: images
          type: array
      required:
      - images
      title: ListImagesResponseContent
      type: object
    ListOfficialImagesResponseContent:
      example:
        images:
        - amiId: amiId
          os: os
          name: name
          version: version
          architecture: architecture
        - amiId: amiId
          os: os
          name: name
          version: version
          architecture: architecture
      properties:
        images:
          items:
            $ref: '#/components/schemas/AmiInfo'
          title: images
          type: array
      required:
      - images
      title: ListOfficialImagesResponseContent
      type: object
...
    NotFoundExceptionResponseContent:
      description: This exception is thrown when the requested entity is not found.
      properties:
        message:
          title: message
          type: string
      title: NotFoundExceptionResponseContent
      type: object
...
    Tag:
      example:
        value: value
        key: key
      properties:
        key:
          description: Tag name
          title: key
          type: string
        value:
          description: Tag value
          title: value
          type: string
      title: Tag
      type: object
    UnauthorizedClientErrorResponseContent:
      description: This exception is thrown when the client is not authorized to perform
        an action.
      properties:
        message:
          title: message
          type: string
      title: UnauthorizedClientErrorResponseContent
      type: object
...
    ValidationLevel:
      enum:
      - INFO
      - WARNING
      - ERROR
      title: ValidationLevel
      type: string
  securitySchemes:
    aws.auth.sigv4:
      description: AWS Signature Version 4 authentication
      in: header
      name: Authorization
      type: apiKey
      x-amazon-apigateway-authtype: awsSigv4
      x-apikeyInfoFunc: pcluster.api.controllers.security_controller_.info_from_aws.auth.sigv4